# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/event_dates": {
            "get": {
                "description": "List Event Dates\nGet a paginated list of ISO dates (e.g. yyyy-mm-dd) for which events exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dates",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid pagination parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/event_entity_ids": {
            "get": {
                "description": "List Event Entity IDs\nGet a paginated list of entity IDs for which events exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity IDs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid pagination parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/event_entity_types": {
            "get": {
                "description": "List Event Entity Types\nGet a complete, sorted list of entity types for which events exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity Types",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/event_log_levels": {
            "get": {
                "description": "List Event Log Levels\nGet a complete, sorted list of log levels for which events exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log Levels",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
...
        "/v1/events/{id}": {
            "get": {
                "description": "Get Event\nGet Event by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Event\nDelete and return the specified Event.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted Event",
                        "schema": {
                            "$ref": "#/definitions/event.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "head": {
                "description": "Event Exists\nCheck if the specified Event exists.",
                "tags": [
                    "Event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event Exists"
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "event.Event": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "entityId": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "otherIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uri": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "main.APIEvent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventID": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
    }
}